Separate CustomActionDefinition from CustomAction
