+++ src/words/environment/WordsClass.java
@@ -4,14 +4,14 @@ import java.util.HashMap;
 
 public class WordsClass {
 	private HashMap<String, Property> properties;
-	private HashMap<String, CustomAction> customActions;
+	private HashMap<String, CustomActionDefinition> customActions;
 	private WordsClass parent;
 	private String className;
 	private ArrayList<WordsClass> children;
 	
 	public WordsClass(String className, WordsClass parent) {
 		properties = new HashMap<String, Property>();
-		customActions = new HashMap<String, CustomAction>();
+		customActions = new HashMap<String, CustomActionDefinition>();
 		children = new ArrayList<WordsClass>();
 		this.parent = parent;
 		this.className = className;
@@ -81,12 +81,12 @@ public class WordsClass {
 	}
 	
 	
-	public void addCustomAction(String customActionName, CustomAction customAction) {
+	public void defineCustomAction(String customActionName, CustomActionDefinition customAction) {
 		//TODO: ActionAlreadyDefined exception
 		customActions.put(customActionName, customAction);
 	}
 	
-	public CustomAction getCustomAction(String customActionName) {
+	public CustomActionDefinition getCustomActionDefinition(String customActionName) {
 		//TODO: throw an exception if the action is not found
 		//TODO: inheritence
 		return customActions.get(customActionName);
